screenshot: long filename handling improvements
For screenshot filenames, it was possible for the basename to be
longer than what filesystems generally support.

On Linux, this is 255 bytes. On Windows, this is 255 wchar_t units.
Thus basenames are truncated to under 255 bytes so that the
basename + extension are <= 255.

The function `trim_invalid_utf8` trims a trailing UTF-8 codepoint
in the case that the string truncation cut it off early.

For Windows, `fopen` will still fail if the file path
is longer than `MAX_PATH` (260). This is worked around, at least on
Windows 10, by prefixing the path with the long-path prefix `\\?\`.

More information on long paths and `\\?\` can be found here:
https://docs.microsoft.com/en-us/windows/win32/fileio/naming-a-file
https://docs.microsoft.com/en-us/windows/win32/fileio/maximum-file-path-limitation

For testing, filling `screenshot-template=` with 3-byte or 4-byte
UTF-8 codepoints is best. Such as "ウ" (3-byte) or "🌂" (4-byte).
Example: 84 * strlen("ウ") + strlen(".jpg") == 256
The last "ウ" is removed and the basename string will be
filled with 83 "ウ" characters and ".jpg" totalling 253 bytes.