screenshot: long filename handling improvements
For screenshot filenames, it was possible for the basename to be
longer than what filesystems generally support.

On Linux, this is 255 bytes. On Windows, this is 255 wchar_t units.
Thus basenames are truncated to under 255 bytes so that the
basename + extension are <= 255 with `truncate_long_base_filename`.
It also makes sure not to produce an invalid UTF-8 codepoint in the filename.

For testing, filling `screenshot-template=` with 3-byte or 4-byte
UTF-8 codepoints is best. Such as "ウ" (3-byte) or "🌂" (4-byte).
Example: 84 * strlen("ウ") + strlen(".jpg") == 256
The last "ウ" is removed and the basename string will be
filled with 83 "ウ" characters and ".jpg" totalling 253 bytes.